Organize worker channel tests
